feat: metadata/fonts/scripts in head for app router
Introduces `PageHead` component, which renders title/meta/link data for pages in
both app and pages router. Currently only supports snippet rendering for pages
router. For pages router, uses Next/Head to render content directly in document
head. For app router, uses React canary extensions to DOM elements, which are
automatically rendered in the `head` tag.
