+++ .changeset/few-scissors-sleep.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+Introduces PageHead component to the base Makeswift Page. This component renders head tag data (link/title/meta) for pages in both app router and pages router. Currently does not support snippets for app router.
