+++ packages/runtime/src/next/hooks/use-is-pages-router.ts
@@ -1,0 +1,10 @@
+import { useRouter } from 'next/router'
+
+export function useIsPagesRouter() {
+  try {
+    useRouter()
+    return true
+  } catch (error) {
+    return false
+  }
+}
